move shout functions to character.cpp
